servicios/servicio-citas/requirements.txt
fastapi>=0.119.1
uvicorn>=0.34.0
pydantic>=2.12.3
pymongo>=4.10.0
python-dotenv>=1.0.1